Desktop/goodbooks-rag-chatbot-main/requirements.txt
streamlit==1.39.0
langchain==0.2.7
langchain-community==0.2.6
langchain-core==0.2.12
langchain-openai==0.1.8
langchain-text-splitters==0.2.2
openai==1.109.1
faiss-cpu==1.8.0
tiktoken==0.7.0
python-dotenv==1.0.1
